диплом. основной сервис. 2 проверка

+++ ewm-service/src/main/java/ru/yandex/practicum/events/dto/EventFullDto.java
@@ -35,6 +35,6 @@ public class EventFullDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime publishedOn;
     private State state;
-    private Long views;
+    private int views;
     private int confirmedRequests;
 }
